requirements.txt
tensorflow-cpu
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
opencv-python
streamlit==1.15.2
